% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonipat_village.R
\docType{data}
\name{sonipat_village}
\alias{sonipat_village}
\title{Survey of India Sonipat village boundaries}
\format{
## `sonipat_village`
Simple feature collection with 324 features and 6 fields:
\describe{
  \item{n_v_code}{Village Code}
  \item{n_v_name}{Name of the Village}
  \item{n_t_code}{Tehsil Code}
  \item{n_t_name}{Tehsil Name}
  \item{n_d_code}{District Code}
  \item{n_d_name}(District Name)
  \item{geom}{Geometry}
}
}
\source{
Survey of India
}
\usage{
sonipat_village
}
\description{
A Sample dataset for Master students at CUH Mahendragarh
}
\keyword{datasets}
